# test: speed up the keyboard tests