## Description

This PR sets the scroll duration to `0` for the date-picker keyboard tests to speed them up.

**Before:**

```
yarn run v1.22.18
$ web-test-runner --group date-picker

Finished running tests in 14.2s, all tests passed! 🎉

✨  Done in 14.67s.
```

**After:**

```
yarn run v1.22.18
$ web-test-runner --group date-picker

Finished running tests in 4.9s, all tests passed! 🎉

✨  Done in 5.28s.
```

## Type of change

- [x] Internal

## Checklist

- [x] I have read the contribution guide: https://vaadin.com/docs-beta/latest/guide/contributing/overview/
- [x] I have added a description following the guideline.
- [x] The issue is created in the corresponding repository and I have referenced it.
- [x] I have added tests to ensure my change is effective and works as intended.
- [x] New and existing tests are passing locally with my change.
- [ ] I have performed self-review and corrected misspellings.